Extract region worker helpers
