+++ fiberhmm/inference/parallel.py
@@ -23,11 +23,6 @@ from fiberhmm.inference.engine import (
     _process_single_read,
     make_apply_payload,
 )
-from fiberhmm.inference.fused_stages import (
-    apply_result_has_footprints,
-    build_fused_recall_result,
-    run_hmm_apply_stage,
-)
 from fiberhmm.inference.read_filters import ReadFilterConfig, streaming_skip_reason
 from fiberhmm.inference.region_planning import (
     _get_genome_regions,
@@ -43,6 +38,13 @@ from fiberhmm.inference.region_types import (
     RegionBedResult,
     RegionBedWorkItem,
 )
+from fiberhmm.inference.region_workers import (
+    _init_fused_region_worker,
+    _init_region_worker,
+    _process_region_to_bam,
+    _process_region_to_bam_fused,
+    _process_region_to_bed,
+)
 from fiberhmm.inference.streaming_drain import (
     _drain_oldest_chunk,
     _drain_oldest_fused_chunk,
@@ -56,15 +58,13 @@ from fiberhmm.inference.streaming_workers import (
 )
 from fiberhmm.inference.tagging import (
     set_legacy_apply_tags,
-    write_fused_recall_tags,
 )
 from fiberhmm.inference.worker_results import coerce_worker_chunk_result
 from fiberhmm.posteriors.region_tsv import (
-    format_region_posterior_line,
-    region_posteriors_tsv_output_path,
+    merge_region_posteriors_tsv as _merge_region_posteriors_tsv,
 )
 from fiberhmm.posteriors.region_tsv import (
-    merge_region_posteriors_tsv as _merge_region_posteriors_tsv,
+    region_posteriors_tsv_output_path,
 )
 
 _is_main_chromosome = _region_is_main_chromosome
@@ -111,398 +111,6 @@ def _select_mp_context() -> 'multiprocessing.context.BaseContext':
 
 
 _MP_CONTEXT = _select_mp_context()
-
-
-# Global for worker processes
-_worker_model = None
-_worker_region_params = None
-_worker_recall_state = {}
-
-
-def _init_region_worker(model_path: str, params: dict):
-    """Initialize worker for region-parallel processing."""
-    global _worker_model, _worker_region_params
-    import os
-
-    import numpy as np
-
-    try:
-        # Disable numba caching to avoid file lock contention
-        os.environ['NUMBA_CACHE_DIR'] = ''
-
-        # Load model once per worker
-        _worker_model = freeze_model_for_inference(load_model(model_path))
-        _worker_region_params = params
-
-        # Warmup numba JIT (just the basic predict, posteriors will warmup on first use)
-        from fiberhmm.core.hmm import HAS_NUMBA
-        if HAS_NUMBA:
-            dummy_obs = np.array([0, 1, 2, 3], dtype=np.int32)
-            _ = _worker_model.predict(dummy_obs)
-
-    except Exception as e:
-        import traceback
-        print(f"Region worker init error: {e}")
-        traceback.print_exc()
-        raise
-
-
-def _process_region_to_bam(args: RegionBamWorkItem) -> RegionBamResult:
-    """
-    Worker function: process one genomic region and write to temp BAM.
-
-    Each worker opens its own BAM file handle and uses the index to fetch
-    reads from its assigned region. This enables true parallel I/O.
-
-    Uses global _worker_model and _worker_region_params (set by _init_region_worker).
-
-    Args:
-        args: RegionBamWorkItem, or the legacy tuple shape.
-
-    Returns:
-        RegionBamResult with temp BAM, counts, optional TSV path, and skip reasons.
-    """
-    import traceback
-
-    global _worker_model, _worker_region_params
-
-    try:
-        work_item = RegionBamWorkItem.from_value(args)
-        chrom, start, end = work_item.region
-        input_bam = work_item.input_bam
-        temp_bam_path = work_item.temp_bam_path
-        temp_tsv_path = work_item.temp_tsv_path
-
-        # Ensure start/end are Python ints (not numpy)
-        start = int(start)
-        end = int(end)
-
-        # Use global model and params (loaded once per worker)
-        model = _worker_model
-        params = _worker_region_params
-
-        # Unpack parameters
-        edge_trim = int(params['edge_trim'])
-        circular = params['circular']
-        mode = params['mode']
-        context_size = int(params['context_size'])
-        msp_min_size = int(params['msp_min_size'])
-        nuc_min_size = int(params.get('nuc_min_size', 85))
-        min_mapq = int(params['min_mapq'])
-        prob_threshold = int(params['prob_threshold'])
-        min_read_length = int(params['min_read_length'])
-        with_scores = params['with_scores']
-        train_rids = params['train_rids']
-        primary_only = params.get('primary_only', False)
-        return_posteriors = params.get('return_posteriors', False) and temp_tsv_path is not None
-        write_msps = params.get('write_msps', True)
-        io_threads = int(params.get('io_threads', 4))
-
-        total_reads = 0
-        reads_with_footprints = 0
-        written = 0
-        skipped = 0
-        posteriors_written = 0
-
-        # Track skip reasons
-        skip_reasons = {
-            'unmapped': 0,
-            'secondary_supplementary': 0,
-            'low_mapq': 0,
-            'too_short': 0,
-            'training_excluded': 0,
-            'no_modifications': 0,
-            'extraction_failed': 0,
-            'no_footprints': 0,
-        }
-
-        # Suppress htslib warnings
-        pysam.set_verbosity(0)
-
-        # Open posteriors TSV file for streaming writes (if requested)
-        tsv_file = None
-        if return_posteriors and temp_tsv_path:
-            try:
-                tsv_file = open(temp_tsv_path, 'w')
-            except Exception:
-                return_posteriors = False  # Can't write, disable
-
-        try:
-            with pysam.AlignmentFile(input_bam, "rb", threads=io_threads, check_sq=False) as inbam:
-                with pysam.AlignmentFile(temp_bam_path, "wb", header=inbam.header, threads=io_threads) as outbam:
-
-                    # Fetch reads from this region using the index
-                    try:
-                        read_iter = inbam.fetch(chrom, start, end)
-                    except ValueError:
-                        # Region not in BAM (e.g., unplaced contigs)
-                        if tsv_file:
-                            tsv_file.close()
-                        return RegionBamResult(temp_bam_path, 0, 0, 0)
-
-                    for read in read_iter:
-                        # Pass through unmapped reads (no sequence to process)
-                        if read.is_unmapped:
-                            outbam.write(read)
-                            written += 1
-                            skipped += 1
-                            skip_reasons['unmapped'] += 1
-                            continue
-
-                        # Skip secondary/supplementary if primary_only mode
-                        if primary_only and (read.is_secondary or read.is_supplementary):
-                            outbam.write(read)
-                            written += 1
-                            skipped += 1
-                            skip_reasons['secondary_supplementary'] += 1
-                            continue
-
-                        # Only process reads that START in this region to avoid duplicates
-                        # (fetch returns reads that overlap the region)
-                        if read.reference_start < start or read.reference_start >= end:
-                            continue
-
-                        # Check filters
-                        if read.mapping_quality < min_mapq:
-                            outbam.write(read)
-                            written += 1
-                            skipped += 1
-                            skip_reasons['low_mapq'] += 1
-                            continue
-
-                        if read.query_alignment_length is None or read.query_alignment_length < min_read_length:
-                            outbam.write(read)
-                            written += 1
-                            skipped += 1
-                            skip_reasons['too_short'] += 1
-                            continue
-
-                        read_id = read.query_name
-                        if train_rids and read_id in train_rids:
-                            outbam.write(read)
-                            written += 1
-                            skipped += 1
-                            skip_reasons['training_excluded'] += 1
-                            continue
-
-                        # Extract data needed for processing
-                        try:
-                            fiber_read = _extract_fiber_read_from_pysam(read, mode, prob_threshold)
-                            if fiber_read is None:
-                                outbam.write(read)
-                                written += 1
-                                skipped += 1
-                                skip_reasons['no_modifications'] += 1
-                                continue
-                        except Exception:
-                            outbam.write(read)
-                            written += 1
-                            skipped += 1
-                            skip_reasons['extraction_failed'] += 1
-                            continue
-
-                        total_reads += 1
-
-                        # Process the read (with posteriors if requested)
-                        result = _process_single_read(
-                            fiber_read, model, edge_trim, circular,
-                            mode, context_size, msp_min_size, nuc_min_size=nuc_min_size,
-                            with_scores=with_scores,
-                            return_posteriors=return_posteriors
-                        )
-
-                        # Add tags to read
-                        if result is not None:
-                            reads_with_footprints += 1
-
-                            set_legacy_apply_tags(read, result, with_scores, write_msps)
-
-                            # Stream posteriors to TSV immediately (no memory accumulation)
-                            if tsv_file and result.get('posteriors') is not None:
-                                try:
-                                    tsv_file.write(
-                                        format_region_posterior_line(
-                                            read_name=read.query_name,
-                                            chrom=read.reference_name,
-                                            ref_start=read.reference_start,
-                                            ref_end=read.reference_end,
-                                            strand=result.get('strand', '.'),
-                                            posteriors=result['posteriors'],
-                                            footprint_starts=result['ns'],
-                                            footprint_sizes=result['nl'],
-                                        )
-                                    )
-                                    posteriors_written += 1
-                                except Exception:
-                                    pass  # Don't crash on posteriors write failure
-                        else:
-                            skip_reasons['no_footprints'] += 1
-
-                        outbam.write(read)
-                        written += 1
-
-        finally:
-            if tsv_file:
-                tsv_file.close()
-
-        # Return TSV path if we wrote any posteriors
-        if return_posteriors and posteriors_written > 0 and temp_tsv_path:
-            return RegionBamResult(
-                temp_bam_path, total_reads, reads_with_footprints,
-                written, temp_tsv_path, skip_reasons,
-            )
-
-        return RegionBamResult(
-            temp_bam_path, total_reads, reads_with_footprints,
-            written, None, skip_reasons,
-        )
-
-    except Exception as e:
-        import traceback
-        print(f"\nWorker error in region {chrom}:{start}-{end}: {e}")
-        traceback.print_exc()
-        raise
-
-def _process_region_to_bed(args: RegionBedWorkItem) -> RegionBedResult:
-    """
-    Process a genomic region and write BED output directly (no temp BAM).
-
-    This is more space-efficient than _process_region_to_bam when only
-    BED/bigBed output is needed.
-
-    Args:
-        args: RegionBedWorkItem, or the legacy tuple shape.
-
-    Returns:
-        RegionBedResult with temp BED path and counts.
-    """
-    work_item = RegionBedWorkItem.from_value(args)
-    region = work_item.region
-    input_bam = work_item.input_bam
-    temp_bed_path = work_item.temp_bed_path
-    chrom, start, end = region
-
-    try:
-        start = int(start)
-        end = int(end)
-
-        model = _worker_model
-        params = _worker_region_params
-
-        edge_trim = int(params['edge_trim'])
-        circular = params['circular']
-        mode = params['mode']
-        context_size = int(params['context_size'])
-        msp_min_size = int(params['msp_min_size'])
-        nuc_min_size = int(params.get('nuc_min_size', 85))
-        min_mapq = int(params['min_mapq'])
-        prob_threshold = int(params['prob_threshold'])
-        min_read_length = int(params['min_read_length'])
-        with_scores = params['with_scores']
-        train_rids = params['train_rids']
-        io_threads = int(params.get('io_threads', 4))
-
-        total_reads = 0
-        reads_with_footprints = 0
-
-        pysam.set_verbosity(0)
-
-        with pysam.AlignmentFile(input_bam, "rb", threads=io_threads, check_sq=False) as inbam:
-            with open(temp_bed_path, 'w') as bed_out:
-                try:
-                    read_iter = inbam.fetch(chrom, start, end)
-                except ValueError:
-                    return RegionBedResult(temp_bed_path, 0, 0)
-
-                for read in read_iter:
-                    # Skip unmapped/secondary/supplementary
-                    if read.is_unmapped or read.is_secondary or read.is_supplementary:
-                        continue
-
-                    # Only process reads that START in this region
-                    if read.reference_start < start or read.reference_start >= end:
-                        continue
-
-                    if read.mapping_quality < min_mapq:
-                        continue
-
-                    if read.query_alignment_length is None or read.query_alignment_length < min_read_length:
-                        continue
-
-                    read_id = read.query_name
-                    if train_rids and read_id in train_rids:
-                        continue
-
-                    try:
-                        fiber_read = _extract_fiber_read_from_pysam(read, mode, prob_threshold)
-                        if fiber_read is None:
-                            continue
-                    except Exception:
-                        continue
-
-                    total_reads += 1
-
-                    result = _process_single_read(
-                        fiber_read, model, edge_trim, circular,
-                        mode, context_size, msp_min_size, nuc_min_size=nuc_min_size,
-                        with_scores=with_scores
-                    )
-
-                    if result is not None and len(result['ns']) > 0:
-                        reads_with_footprints += 1
-
-                        # Write BED12 line
-                        ref_name = read.reference_name
-                        ref_start = read.reference_start
-                        ref_end = read.reference_end
-                        strand = '-' if read.is_reverse else '+'
-                        read_length = ref_end - ref_start
-
-                        # Footprint blocks (convert to relative positions)
-                        ns = result['ns']
-                        nl = result['nl']
-                        block_starts_list = [int(s - ref_start) for s in ns]
-                        block_sizes_list = [int(length) for length in nl]
-
-                        # Get scores if available
-                        score_list = None
-                        if with_scores and result['ns_scores'] is not None:
-                            score_list = [int(s * 1000) for s in result['ns_scores']]
-
-                        # BED12 requires blocks to span chromStart to chromEnd
-                        # Add 1bp padding blocks at start/end if needed
-                        if block_starts_list[0] != 0:
-                            block_starts_list.insert(0, 0)
-                            block_sizes_list.insert(0, 1)
-                            if score_list is not None:
-                                score_list.insert(0, 0)
-
-                        last_end = block_starts_list[-1] + block_sizes_list[-1]
-                        if last_end < read_length:
-                            block_starts_list.append(read_length - 1)
-                            block_sizes_list.append(1)
-                            if score_list is not None:
-                                score_list.append(0)
-
-                        block_count = len(block_starts_list)
-                        block_sizes = ','.join(str(s) for s in block_sizes_list)
-                        block_starts = ','.join(str(s) for s in block_starts_list)
-
-                        if score_list is not None:
-                            scores = ','.join(str(s) for s in score_list)
-                            bed_out.write(f"{ref_name}\t{ref_start}\t{ref_end}\t{read_id}\t0\t{strand}\t"
-                                        f"{ref_start}\t{ref_end}\t0,0,0\t{block_count}\t{block_sizes}\t{block_starts}\t{scores}\n")
-                        else:
-                            bed_out.write(f"{ref_name}\t{ref_start}\t{ref_end}\t{read_id}\t0\t{strand}\t"
-                                        f"{ref_start}\t{ref_end}\t0,0,0\t{block_count}\t{block_sizes}\t{block_starts}\n")
-
-        return RegionBedResult(temp_bed_path, total_reads, reads_with_footprints)
-
-    except Exception as e:
-        import traceback
-        print(f"\nWorker error in region {chrom}:{start}-{end}: {e}")
-        traceback.print_exc()
-        raise
 
 
 def _process_bam_region_parallel(input_bam: str, output_bam: str,
@@ -901,229 +509,6 @@ def _process_and_write_chunk(chunk_reads: list, chunk_read_objs: list,
             list(zip(chunk_read_objs, chunk_reads, results)),
         )
     return reads_with_footprints, no_footprints, worker_failures, None
-
-
-def _init_fused_region_worker(apply_model_path: str, recall_model_path: Optional[str],
-                              emission_uplift: float, params: dict):
-    """Per-worker init for region-parallel fused apply+recall.
-
-    Loads the apply HMM model, builds the TF LLR tables (from the recall
-    model or by reusing the apply model), warms up numba JIT, and stashes
-    params for the region worker to pick up.
-    """
-    global _worker_model, _worker_region_params, _worker_recall_state
-    import os
-    os.environ['NUMBA_CACHE_DIR'] = ''
-
-    _worker_model = freeze_model_for_inference(load_model(apply_model_path))
-    # Open the reference FASTA *after* fork: pysam.FastaFile is not
-    # fork-safe. Stash the live handle on params so the region worker
-    # can pass it through to get_daf_positions.
-    ref_path = params.get('ref_fasta_path')
-    if ref_path:
-        import pysam as _pysam
-        params = dict(params)   # don't mutate the shared-across-workers dict
-        params['ref_fasta'] = _pysam.FastaFile(ref_path)
-    _worker_region_params = params
-
-    from fiberhmm.core.model_io import load_model_with_metadata
-    from fiberhmm.inference.tf_recaller import apply_emission_uplift, build_llr_tables
-    r_path = recall_model_path or apply_model_path
-    r_model, _, _ = load_model_with_metadata(r_path)
-    llr_hit, llr_miss = build_llr_tables(r_model)
-    if abs(emission_uplift - 1.0) > 1e-9:
-        llr_hit, llr_miss = apply_emission_uplift(llr_hit, llr_miss, r_model, emission_uplift)
-    _worker_recall_state['llr_hit'] = llr_hit
-    _worker_recall_state['llr_miss'] = llr_miss
-
-    # Warmup
-    from fiberhmm.core.hmm import HAS_NUMBA
-    if HAS_NUMBA:
-        dummy_obs = np.array([0, 1, 2, 3], dtype=np.int32)
-        _ = _worker_model.predict(dummy_obs)
-        from fiberhmm.inference.tf_recaller import call_tfs_in_interval
-        _ = call_tfs_in_interval(
-            np.zeros(16, dtype=np.int32), 0, 16,
-            llr_hit, llr_miss, min_llr=4.0, min_opps=3,
-        )
-
-
-def _process_region_to_bam_fused(args: RegionBamWorkItem) -> RegionBamResult:
-    """Region worker: fetch reads in one genomic region, run fused
-    apply+recall per read, write in-order to a coordinate-sorted temp BAM.
-
-    Because pysam.fetch(chrom,start,end) yields reads in coordinate order
-    AND we only process reads that START in this region (the reference_start
-    filter), each temp BAM is coordinate-sorted within itself.  Concatenating
-    temp BAMs in region order gives a coordinate-sorted final BAM without
-    any sort pass.
-
-    Returns a RegionBamResult with temp BAM, counts, and skip reasons.
-    """
-    import traceback
-
-    from fiberhmm.inference.engine import extract_fiber_read_from_payload, make_apply_payload
-    global _worker_model, _worker_region_params, _worker_recall_state
-
-    try:
-        work_item = RegionBamWorkItem.from_value(args)
-        chrom, start, end = work_item.region
-        input_bam = work_item.input_bam
-        temp_bam_path = work_item.temp_bam_path
-        start = int(start)
-        end = int(end)
-
-        params = _worker_region_params
-        model = _worker_model
-        llr_hit = _worker_recall_state['llr_hit']
-        llr_miss = _worker_recall_state['llr_miss']
-
-        edge_trim = int(params['edge_trim'])
-        circular = params['circular']
-        mode = params['mode']
-        ref_fasta = params.get('ref_fasta')   # opened FastaFile or None; --reference path
-        context_size = int(params['context_size'])
-        msp_min_size = int(params['msp_min_size'])
-        nuc_min_size = int(params.get('nuc_min_size', 85))
-        min_mapq = int(params['min_mapq'])
-        prob_threshold = int(params['prob_threshold'])
-        min_read_length = int(params['min_read_length'])
-        with_scores = params.get('with_scores', False)
-        train_rids = params.get('train_rids') or set()
-        primary_only = params.get('primary_only', False)
-        io_threads = int(params.get('io_threads', 4))
-        min_llr = float(params['min_llr'])
-        min_opps = int(params['min_opps'])
-        unify_threshold = int(params['unify_threshold'])
-        also_write_legacy = params['also_write_legacy']
-        downstream_compat = params['downstream_compat']
-
-        pysam.set_verbosity(0)
-
-        total_reads = 0
-        reads_with_fp = 0
-        written = 0
-        skipped = 0
-        skip_reasons = {
-            'unmapped': 0, 'secondary_supplementary': 0, 'low_mapq': 0,
-            'too_short': 0, 'training_excluded': 0, 'no_modifications': 0,
-            'extraction_failed': 0, 'no_footprints': 0,
-        }
-
-        with pysam.AlignmentFile(input_bam, "rb", threads=io_threads, check_sq=False) as inbam:
-            with pysam.AlignmentFile(temp_bam_path, "wb", header=inbam.header, threads=io_threads) as outbam:
-                try:
-                    read_iter = inbam.fetch(chrom, start, end)
-                except ValueError:
-                    return RegionBamResult(temp_bam_path, 0, 0, 0)
-
-                for read in read_iter:
-                    if read.is_unmapped:
-                        outbam.write(read)
-                        written += 1
-                        skipped += 1
-                        skip_reasons['unmapped'] += 1
-                        continue
-                    if primary_only and (read.is_secondary or read.is_supplementary):
-                        outbam.write(read)
-                        written += 1
-                        skipped += 1
-                        skip_reasons['secondary_supplementary'] += 1
-                        continue
-                    # Only process reads starting in this region (fetch is overlap-based).
-                    if read.reference_start < start or read.reference_start >= end:
-                        continue
-                    if read.mapping_quality < min_mapq:
-                        outbam.write(read)
-                        written += 1
-                        skipped += 1
-                        skip_reasons['low_mapq'] += 1
-                        continue
-                    if (read.query_alignment_length is None
-                            or read.query_alignment_length < min_read_length):
-                        outbam.write(read)
-                        written += 1
-                        skipped += 1
-                        skip_reasons['too_short'] += 1
-                        continue
-                    if train_rids and read.query_name in train_rids:
-                        outbam.write(read)
-                        written += 1
-                        skipped += 1
-                        skip_reasons['training_excluded'] += 1
-                        continue
-
-                    payload = make_apply_payload(read, mode=mode, ref_fasta=ref_fasta)
-                    if payload is None:
-                        outbam.write(read)
-                        written += 1
-                        skipped += 1
-                        skip_reasons['no_modifications'] += 1
-                        continue
-
-                    try:
-                        fiber_read = extract_fiber_read_from_payload(payload, mode, prob_threshold)
-                        if fiber_read is None:
-                            outbam.write(read)
-                            written += 1
-                            skipped += 1
-                            skip_reasons['no_modifications'] += 1
-                            continue
-                        apply_result = run_hmm_apply_stage(
-                            fiber_read,
-                            model,
-                            edge_trim,
-                            circular,
-                            mode,
-                            context_size,
-                            msp_min_size,
-                            nuc_min_size,
-                            with_scores,
-                        )
-                    except Exception:
-                        outbam.write(read)
-                        written += 1
-                        skipped += 1
-                        skip_reasons['extraction_failed'] += 1
-                        continue
-
-                    total_reads += 1
-
-                    if not apply_result_has_footprints(apply_result):
-                        outbam.write(read)
-                        written += 1
-                        skip_reasons['no_footprints'] += 1
-                        continue
-
-                    fused_result = build_fused_recall_result(
-                        fiber_read,
-                        apply_result,
-                        llr_hit,
-                        llr_miss,
-                        min_llr,
-                        min_opps,
-                        unify_threshold,
-                        with_scores,
-                    )
-                    write_fused_recall_tags(
-                        read,
-                        read_length=len(fiber_read['query_sequence']),
-                        result=fused_result,
-                        also_write_legacy=also_write_legacy,
-                        downstream_compat=downstream_compat,
-                    )
-                    outbam.write(read)
-                    written += 1
-                    reads_with_fp += 1
-
-        return RegionBamResult(
-            temp_bam_path, total_reads, reads_with_fp,
-            written, None, skip_reasons,
-        )
-
-    except Exception:
-        traceback.print_exc()
-        raise
 
 
 def _process_bam_region_parallel_fused(
